I've confirmed all numbers are accurate with the exception of frequency response. I don't fully understand why this number is as high as it is.
All of the numbers that were calculated are slightly different from what we did by hand but this makes sense as the hand calc. assumed some perfect numbers. When doing it programmatically, we're actually calc. these numbers so they're slightly different. For example We assumed CCAB was 7E-7 but calc. shows it's actually a little off from that. This leads Ral to go from a perfect 90k to 89446.68. Things continue to snowball from there. However, the values are still relatively close.

Signed-off-by: Damian <damianhgn@gmail.com>